Compiling but not functional yet.
I believe the problem with the program now is that it reads over the data many times per block, instead of just once. My suspicion is that there is a problem when reading the decompressed data into a string. The loop iterator, have, is always the block size, meaning each loop iteration parses the entire block. This would be fine if its while loop (dictated by strm.avail_in != 0) ran just once but it doesn't.