Worked on FunctionCall test.
Are we getting those bad global stores because there are instructions that are tripping the condition to create stores in GetMemoryFunctions? There is a key value pair that indicates all global values have their own instruction. At this point, probably yes. The alloca instructions being put into the same scope may be pointing at the same input args and the same data types. Wouldn't that match? Also a solution could be to get rid of alloca's entirely and replace them with our own instructions for input args.